refactor code, centralizing data retrieval with CacheFactory
