+++ src/app/app.constants.ts
@@ -2,7 +2,5 @@ import { Injectable } from "@angular/core";
 
 @Injectable()
 export class AppConstants { 
-    readonly GIT_URL: string = 'https://raw.githubusercontent.com/GSA/marketplace-fedramp-gov-data/main/data.json';
-    readonly CACHE_DELAY: number = 1000;
-    readonly CACHE_DATE_FORMAT: string = 'yyyy-MM-dd';
+    public static readonly GIT_URL: string = 'https://raw.githubusercontent.com/GSA/marketplace-fedramp-gov-data/main/data.json';
 }
