+++ src/app/product-info/product-info.component.html
@@ -1,4 +1,4 @@
-<div *ngIf="isCachePresent">
+<div *ngIf="cf.isCachePresent()">
   <div id="product-info" style="display:none;"></div>
 
   <div id="update-title" style="display:none;">{{product.cso}}</div>
