+++ src/app/refresh-data/refresh-data.component.html
@@ -1,3 +1,3 @@
 
 <p>To refresh of the most recent data, click below.</p>
-<button id="refresh-button" class="btn btn-primary" onclick="localStorage.removeItem('cacheProducts');localStorage.removeItem('cacheDate'); localStorage.removeItem('cacheAgencies');localStorage.removeItem('cacheAssessors');localStorage.removeItem('cacheFilters');localStorage.removeItem('cacheMetrics');localStorage.removeItem('cacheAtoMapping');localStorage.removeItem('cacheReuseMapping');window.location.href='/';">Refresh</button>
+<button id="refresh-button" class="btn btn-primary" onclick="window.location.href='/';">Refresh</button>
