+++ src/app/refresh-data/refresh-data.component.ts
@@ -7,11 +7,16 @@ import { Component, OnInit } from '@angular/core';
 })
 
 export class RefreshDataComponent implements OnInit {
-
-  constructor() {    
+  constructor() {  
+    localStorage.removeItem("cacheJson");
+    localStorage.removeItem("cacheDate");
+    localStorage.removeItem("cacheProducts");
+    localStorage.removeItem("cacheAgencies");
+    localStorage.removeItem("cacheAssessors");
+    localStorage.removeItem("cacheFilters");
+    localStorage.removeItem("cacheMetrics");
+    localStorage.removeItem("cacheAtoMapping");
+    localStorage.removeItem("cacheReuseMapping");  
   }
-
-  ngOnInit(): void {    
-  }
- 
+  ngOnInit(): void {}
 }
