+++ src/index.html
@@ -6,7 +6,6 @@
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="./assets/images/favicon.ico">
-  <script src="/assets/js/csv.js"></script>
   <script src="/assets/js/filter.js"></script>
   <script src="/assets/js/sort.js"></script>
   <script src="/assets/js/util.js"></script>
